+++ src/ApiManager/types.ts
@@ -1,3 +1,5 @@
+import type { ETracksDirection } from './constants';
+
 export enum ECMDNotify {
   CHANNELS = 'channels',
   WEBCAST_STARTED = 'WebcastStarted',
@@ -110,4 +112,10 @@ export type TOptionsInfoMediaState = {
 
 export type TOptionsExtraHeaders = {
   extraHeaders?: string[];
+};
+
+export type TRestartData = {
+  tracksDirection: ETracksDirection;
+  audioTrackCount: number;
+  videoTrackCount: number;
 };