+++ src/CallManager/@CallManager.ts
@@ -85,8 +85,20 @@ class CallManager {
     return this.strategy.getRemoteStreams();
   };
 
+  public getTransceivers: ICallStrategy['getTransceivers'] = () => {
+    return this.strategy.getTransceivers();
+  };
+
+  public addTransceiver: ICallStrategy['addTransceiver'] = async (...args) => {
+    return this.strategy.addTransceiver(...args);
+  };
+
   public replaceMediaStream: ICallStrategy['replaceMediaStream'] = async (...args) => {
     return this.strategy.replaceMediaStream(...args);
+  };
+
+  public restartIce: ICallStrategy['restartIce'] = async (options) => {
+    return this.strategy.restartIce(options);
   };
 }